# [pkg/stanza] Localize file offset initialization