--- a/pkg/stanza/fileconsumer/reader_factory.go
+++ b/pkg/stanza/fileconsumer/reader_factory.go
@@ -25,6 +25,7 @@ import (
 type readerFactory struct {
 	*zap.SugaredLogger
 	readerConfig   *readerConfig
+	fromBeginning  bool
 	splitterConfig helper.SplitterConfig
 }
 
@@ -113,6 +114,10 @@ func (b *readerBuilder) build() (r *Reader, err error) {
 
 	if b.fp != nil {
 		r.Fingerprint = b.fp
+	}
+
+	if !b.fromBeginning {
+		r.offsetToEnd()
 	}
 
 	return r, nil