This refactoring transfers responsibility for Reader offset initialization to the readerFactory.

Users can configure whether or not to start reading from the beginning or end of files, but this only applies to the initial set of files found in the matching pattern. New files are consumed entirely, so reading from the beginning is assumed after the first poll cycle. 

The boolean that tracks the state of where to start reading, is overridden automatically. This flag does not require a mutex because it is written to at a point in the poll cycle where there is only a single goroutine running.