@djaglowski 2022-07-12
Some related load test failures:
```
Log10kDPS/filelog                       |FAIL  |      3s|    31.3|    31.3|         28|         84|     31800|         29000|CPU consumption is 31.3%, max expected is 30%
Log10kDPS/filelog_checkpoints           |FAIL  |      9s|    30.7|    32.0|         52|         89|     91100|         88900|CPU consumption is 32.0%, max expected is 30%
```

I'm not seeing notable differences in the benchmarks between this and main though.

`main`:
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer
BenchmarkFileInput/Single-10         	  957364	      1282 ns/op	    4120 B/op	       2 allocs/op
BenchmarkFileInput/Glob-10           	  187180	      5640 ns/op	   16481 B/op	       8 allocs/op
BenchmarkFileInput/MultiGlob-10      	  191376	      5662 ns/op	   16480 B/op	       8 allocs/op
BenchmarkFileInput/MaxConcurrent-10  	  149456	      6826 ns/op	   16482 B/op	       8 allocs/op
BenchmarkFileInput/FngrPrntLarge-10  	  978186	      1031 ns/op	    4120 B/op	       2 allocs/op
BenchmarkFileInput/FngrPrntSmall-10  	  985405	      1228 ns/op	    4120 B/op	       2 allocs/op
PASS
coverage: 75.4% of statements
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer	52.662s
```

This branch
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer
BenchmarkFileInput/Single-10         	  967011	      1242 ns/op	    4120 B/op	       2 allocs/op
BenchmarkFileInput/Glob-10           	  188626	      5592 ns/op	   16481 B/op	       8 allocs/op
BenchmarkFileInput/MultiGlob-10      	  191498	      5540 ns/op	   16481 B/op	       8 allocs/op
BenchmarkFileInput/MaxConcurrent-10  	  146421	      6953 ns/op	   16482 B/op	       8 allocs/op
BenchmarkFileInput/FngrPrntLarge-10  	  968253	      1264 ns/op	    4120 B/op	       2 allocs/op
BenchmarkFileInput/FngrPrntSmall-10  	  977587	      1068 ns/op	    4120 B/op	       2 allocs/op
PASS
coverage: 74.8% of statements
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer	55.536s

```

I also see that the [prior run](https://github.com/open-telemetry/opentelemetry-collector-contrib/runs/7308373235?check_suite_focus=true#step:14:316) on this PR passed quite easily:
```
Log10kDPS/filelog                       |PASS  |     15s|    24.3|    25.0|         59|         90|    150000|        150000|
Log10kDPS/filelog_checkpoints           |PASS  |     15s|    23.9|    24.7|         58|         88|    150000|        150000|
```

My inclination is to just raise these limits, as the limits appears to just be a bit too tight.